requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.35.1
polars-runtime-32==1.35.1
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.13.1
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
ultralytics==8.3.224
ultralytics-thop==2.0.18
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
